deployment/backend/requirements.txt
fastapi==0.100.0
pydantic==1.10.7
python-dotenv==1.0.0
pytorch_lightning==2.0.5
torch==2.0.1
torchtext==0.15.2
uvicorn==0.23.0
gunicorn==20.1.0
websockets==11.0.3